+++ CHANGELOG.md
@@ -20,10 +20,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added roles and authorization validation. - [#107](https://github.com/DigitalExcellence/dex-backend/issues/107)
 - Added unittests for UserRepository - [#121] (https://github.com/DigitalExcellence/dex-backend/issues/121)
 - Added unittests for UserService - [#121] (https://github.com/DigitalExcellence/dex-backend/issues/121)
+- Added an endpoint to get information about the current user - [#141](https://github.com/DigitalExcellence/dex-backend/issues/141)
 
 ### Changed
 
 - Return Unauthorized instead of Bad Request when not allowed to perform action in controller - [#132](https://github.com/DigitalExcellence/dex-backend/issues/132)
+- Changed the migrations and seeding of the data - [#134](https://github.com/DigitalExcellence/dex-backend/issues/134)
 - Get user from the session & add current user to project. - [#139](https://github.com/DigitalExcellence/dex-backend/issues/139)
 - Changed Student reference to be named identity. - [#145](https://github.com/DigitalExcellence/dex-backend/issues/145)
 
@@ -36,4 +38,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
-